requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
groq==0.4.0
pymysql==1.1.0
aws-wsgi==0.2.7
boto3==1.34.50